fix: add tests for new data soure
