+++ projects/observability/src/shared/dashboard/data/graphql/percentage/percentage-data-source.model.ts
