+++ projects/observability/src/shared/dashboard/data/graphql/percentage/percentage-data-source.module.ts
